requirements.txt
pygame-ce==2.3.2
opensimplex
numpy
pygame_gui